feat(export): fix QGIS DLL import on Windows Py3.12 via add_dll_directory; isolate worker in modules/export_worker.py; keep parallel export with spawn; add sequential fallback
